EMS motor control work in progress.
svn path=/trunk/iCub/firmware/emBODY/arch-armcmx/body/embobj/embobj/; revision=18866

+++ emBODY/eBcode/arch-arm/embobj/embobj-src/EOtrajectory_hid.h
@@ -53,8 +53,8 @@ struct EOtrajectory_hid
     int32_t pos_steps_to_end;
     int32_t vel_steps_to_end;
 
-    int32_t pos_min;
-    int32_t pos_max;
+    float pos_min;
+    float pos_max;
 };
